commit day 10

+++ db/delete_product.sql
@@ -1,3 +1,3 @@
-DELETE FROM toBuy WHERE list_id = $1;
+DELETE FROM toBuy WHERE list_id = $1 AND user_id = $2;
 
-SELECT * FROM toBuy
+SELECT * FROM toBuy WHERE user_id = $2
